refactor(receipts): decompose ReceiptsService into Issuance/Void/Query/Pdf + ReceiptNumber services
Behavior-preserving regulated money-path. Facade keeps 8-method surface + 4-arg
ctor + delegates; internally constructs sub-services (spec `new` site + module +
PaymentsService consumer untouched). voidReceipt reversal-JE $tx + generateReceipt
issuance $tx moved WHOLE; ~280-LOC PDF render lifted to ReceiptPdfService;
generateReceiptNumber shared (tx-aware). tsc 0, receipts tests green, lint 0.
